requirements.txt
asgiref==3.10.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==5.2.7
django-allauth==65.12.1
djangorestframework==3.16.1
idna==3.11
jwt==1.4.0
pycparser==2.23
pyngrok==7.4.0
python-dotenv==1.1.1
PyYAML==6.0.3
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
